requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Cython==0.28.2
dataclasses==0.5
Django==2.0.4
django-autoslug==1.9.3
django-autoslug-iplweb==1.9.4.dev0
django-cities==0.5.0.6
django-cities-light==3.4.0
django-model-utils==3.1.1
django-rest-swagger==2.1.2
djangorestframework==3.8.2
djongo==1.2.24
ez-setup==0.9
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.2
openapi-codec==1.3.2
progressbar2==3.37.1
pymongo==3.6.1
pystan==2.17.1.0
python-utils==2.3.0
pytz==2018.4
requests==2.18.4
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
swapper==1.1.0
tqdm==4.23.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
